Split @docusaurus/theme-common into public/internal apis
