+++ packages/docusaurus-theme-classic/src/theme/DocItem/Content/index.tsx
@@ -7,7 +7,8 @@
 
 import React from 'react';
 import clsx from 'clsx';
-import {ThemeClassNames, useDoc} from '@docusaurus/theme-common';
+import {ThemeClassNames} from '@docusaurus/theme-common';
+import {useDoc} from '@docusaurus/theme-common/internal';
 import Heading from '@theme/Heading';
 import MDXContent from '@theme/MDXContent';
 import type {Props} from '@theme/DocItem/Content';
