+++ packages/docusaurus-theme-classic/src/theme/DocItem/Layout/index.tsx
@@ -7,7 +7,8 @@
 
 import React from 'react';
 import clsx from 'clsx';
-import {useWindowSize, useDoc} from '@docusaurus/theme-common';
+import {useWindowSize} from '@docusaurus/theme-common';
+import {useDoc} from '@docusaurus/theme-common/internal';
 import DocItemPaginator from '@theme/DocItem/Paginator';
 import DocVersionBanner from '@theme/DocVersionBanner';
 import DocVersionBadge from '@theme/DocVersionBadge';
