+++ packages/docusaurus-theme-classic/src/theme/DocItem/Paginator/index.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import {useDoc} from '@docusaurus/theme-common';
+import {useDoc} from '@docusaurus/theme-common/internal';
 import DocPaginator from '@theme/DocPaginator';
 
 /**
