+++ packages/docusaurus-theme-classic/src/theme/DocItem/TOC/Mobile/index.tsx
@@ -7,7 +7,9 @@
 
 import React from 'react';
 import clsx from 'clsx';
-import {ThemeClassNames, useDoc} from '@docusaurus/theme-common';
+import {ThemeClassNames} from '@docusaurus/theme-common';
+import {useDoc} from '@docusaurus/theme-common/internal';
+
 import TOCCollapsible from '@theme/TOCCollapsible';
 
 import styles from './styles.module.css';
