+++ packages/docusaurus-theme-classic/src/theme/DocItem/index.tsx
@@ -6,7 +6,8 @@
  */
 
 import React from 'react';
-import {HtmlClassNameProvider, DocProvider} from '@docusaurus/theme-common';
+import {HtmlClassNameProvider} from '@docusaurus/theme-common';
+import {DocProvider} from '@docusaurus/theme-common/internal';
 import DocItemMetadata from '@theme/DocItem/Metadata';
 import DocItemLayout from '@theme/DocItem/Layout';
 import type {Props} from '@theme/DocItem';
