+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/Main/index.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import clsx from 'clsx';
-import {useDocsSidebar} from '@docusaurus/theme-common';
+import {useDocsSidebar} from '@docusaurus/theme-common/internal';
 import type {Props} from '@theme/DocPage/Layout/Main';
 
 import styles from './styles.module.css';
