+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/index.tsx
@@ -6,7 +6,7 @@
  */
 
 import React, {useState} from 'react';
-import {useDocsSidebar} from '@docusaurus/theme-common';
+import {useDocsSidebar} from '@docusaurus/theme-common/internal';
 import Layout from '@theme/Layout';
 import BackToTopButton from '@theme/BackToTopButton';
 import DocPageLayoutSidebar from '@theme/DocPage/Layout/Sidebar';
