+++ packages/docusaurus-theme-classic/src/theme/DocSidebarItem/Category/index.tsx
@@ -7,17 +7,16 @@
 
 import React, {type ComponentProps, useEffect, useMemo} from 'react';
 import clsx from 'clsx';
+import {ThemeClassNames, useThemeConfig} from '@docusaurus/theme-common';
 import {
   isActiveSidebarItem,
   usePrevious,
   Collapsible,
   useCollapsible,
   findFirstCategoryLink,
-  ThemeClassNames,
-  useThemeConfig,
   useDocSidebarItemsExpandedState,
   isSamePath,
-} from '@docusaurus/theme-common';
+} from '@docusaurus/theme-common/internal';
 import Link from '@docusaurus/Link';
 import {translate} from '@docusaurus/Translate';
 import useIsBrowser from '@docusaurus/useIsBrowser';
