+++ packages/docusaurus-theme-classic/src/theme/DocVersionBanner/index.tsx
@@ -15,12 +15,11 @@ import {
   useDocVersionSuggestions,
   type GlobalVersion,
 } from '@docusaurus/plugin-content-docs/client';
+import {ThemeClassNames} from '@docusaurus/theme-common';
 import {
-  ThemeClassNames,
   useDocsPreferredVersion,
   useDocsVersion,
-} from '@docusaurus/theme-common';
-
+} from '@docusaurus/theme-common/internal';
 import type {Props} from '@theme/DocVersionBanner';
 import type {
   VersionBanner,
