+++ packages/docusaurus-theme-classic/src/theme/Navbar/MobileSidebar/index.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import {
   useLockBodyScroll,
   useNavbarMobileSidebar,
-} from '@docusaurus/theme-common';
+} from '@docusaurus/theme-common/internal';
 import NavbarMobileSidebarLayout from '@theme/Navbar/MobileSidebar/Layout';
 import NavbarMobileSidebarHeader from '@theme/Navbar/MobileSidebar/Header';
 import NavbarMobileSidebarPrimaryMenu from '@theme/Navbar/MobileSidebar/PrimaryMenu';
