+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DocsVersionNavbarItem.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import {useDocsVersionCandidates} from '@docusaurus/theme-common';
+import {useDocsVersionCandidates} from '@docusaurus/theme-common/internal';
 import DefaultNavbarItem from '@theme/NavbarItem/DefaultNavbarItem';
 import type {Props} from '@theme/NavbarItem/DocsVersionNavbarItem';
 import type {GlobalVersion} from '@docusaurus/plugin-content-docs/client';
