+++ packages/docusaurus-theme-classic/src/theme/NavbarItem/DropdownNavbarItem.tsx
@@ -7,13 +7,13 @@
 
 import React, {useState, useRef, useEffect} from 'react';
 import clsx from 'clsx';
+import {isRegexpStringMatch} from '@docusaurus/theme-common';
 import {
   isSamePath,
   useCollapsible,
   Collapsible,
-  isRegexpStringMatch,
   useLocalPathname,
-} from '@docusaurus/theme-common';
+} from '@docusaurus/theme-common/internal';
 import NavbarNavLink from '@theme/NavbarItem/NavbarNavLink';
 import NavbarItem, {type LinkLikeNavbarItemProps} from '@theme/NavbarItem';
 import type {
