+++ packages/docusaurus-theme-classic/src/theme/SiteMetadata/index.tsx
@@ -9,13 +9,12 @@ import React from 'react';
 import Head from '@docusaurus/Head';
 import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
 import useBaseUrl from '@docusaurus/useBaseUrl';
+import {PageMetadata, useThemeConfig} from '@docusaurus/theme-common';
 import {
-  PageMetadata,
   DEFAULT_SEARCH_TAG,
   useAlternatePageUtils,
-  useThemeConfig,
   keyboardFocusedClassName,
-} from '@docusaurus/theme-common';
+} from '@docusaurus/theme-common/internal';
 import {useLocation} from '@docusaurus/router';
 import SearchMetadata from '@theme/SearchMetadata';
